fixed bugs of restaurants-searchBar merge. Merge now functional
