+++ dib/src/pages/MainPage.css
@@ -1,0 +1,4 @@
+.project-name {
+  font-size: 3rem;
+  text-align: center;
+}
