+++ dib/src/pages/MainPage.js
@@ -1,12 +1,15 @@
-// import SearchBarPage from "./SearchBarPage"
+import './MainPage.css'
+import SearchBarPage from "./SearchBarPage"
 import RestList from "../components/RestList"
 
 function MainPage () {
   console.log('main page before jsx')
   return (
     <div>
-      hallo
-     {/* <SearchBarPage/> */}
+      <h1 className="project-name">Drop into Berlin</h1>
+
+     <SearchBarPage/>
+     
      <RestList/>
     </div>
   )
