+++ dib/src/pages/SearchBarPage.js
@@ -1,4 +1,4 @@
-/* import { useContext, useState } from 'react';
+import { useContext, useState } from 'react';
 import SearchBar from '../components/SearchBar';
 import RestContext from "../context/RestContext"
 
@@ -14,22 +14,27 @@ function SearchBarPage() {
         setSearchSuggestions(filteredResults.slice(0, 8));
         console.log ('handled SearchChange')
     };
+
+    //
+
+    /* const displayedSearchSuggestions = searchSuggestions.map(restaurant => (
+        <div key={restaurant.id} className="border p-4 rounded">
+            <img src={restaurant.imageURL} alt={restaurant.name} className="w-full h-48 object-cover rounded" />
+            <h2 className="text-xl font-semibold mt-2">{restaurant.name}</h2>
+            <p className="text-gray-600">{restaurant.description}</p>
+        </div>
+    )) */
+
     console.log('search bar before return')
     
     return (
         <div className="p-4">
             <SearchBar onSearchChange={handleSearchChange} suggestions={searchSuggestions} />
             <div className="mt-4 space-y-4">
-                {searchSuggestions.map(restaurant => (
-                    <div key={restaurant.id} className="border p-4 rounded">
-                        <img src={restaurant.imageURL} alt={restaurant.name} className="w-full h-48 object-cover rounded" />
-                        <h2 className="text-xl font-semibold mt-2">{restaurant.name}</h2>
-                        <p className="text-gray-600">{restaurant.description}</p>
-                    </div>
-                ))}
+                {/* {searchSuggestions ? displayedSearchSuggestions : null} */}
             </div>
         </div>
     );
 }
 
-export default SearchBarPage; */
+export default SearchBarPage;
